♻️ refactor: feature 폴더를 추가하여 폴더 구조를 개선하였습니다!
