+++ src/features/sign_up/api/signUpAPI.ts
@@ -1,4 +1,4 @@
-import { postRequest } from "../../apis/requests";
+import { postRequest } from "../../../api/requests";
 import { SignUpApiUrl, SignUpRequest, SignUpResponse } from "./signUpAPI.type";
 
 export async function postSignUp(signUpFormData: SignUpRequest) {
