+++ src/pages/Todos/Todos.container.tsx
@@ -1,11 +1,11 @@
-import { useTodosList } from "../../hooks/queries/todos/useTodosList";
+import { useTodosList } from "../../features/todolist/queries/useTodosList";
 import TodosPresenter from "./Todos.presenter";
 
 function TodosContainer() {
   const { data: todos } = useTodosList();
 
   if (!todos) {
-    return;
+    return null;
   }
 
   return <TodosPresenter todos={todos} />;
